[#72] 회원 프로필 편집 API

+++ src/main/java/com/poortorich/s3/service/FileUploadService.java
@@ -5,11 +5,13 @@ import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.poortorich.global.exceptions.InternalServerErrorException;
+import com.poortorich.s3.constants.S3Constants;
 import com.poortorich.s3.response.enums.S3Response;
 import com.poortorich.s3.util.S3FileUtils;
 import com.poortorich.s3.validator.FileValidator;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -19,13 +21,18 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class FileUploadService {
 
+
     private final AmazonS3Client amazonS3Client;
     private final FileValidator fileValidator;
-    
+
     @Value("${cloud.aws.s3.bucket}")
     private String bucketName;
 
     public String uploadImage(MultipartFile imageFile) {
+        if (imageFile.isEmpty()) {
+            return S3Constants.DEFAULT_PROFILE_IMAGE;
+        }
+
         fileValidator.validateFileType(imageFile);
         fileValidator.validateFileSize(imageFile);
 
@@ -36,9 +43,25 @@ public class FileUploadService {
         return amazonS3Client.getUrl(bucketName, fileName).toString();
     }
 
+    public String updateImage(String currentProfileImage, MultipartFile newProfileImage, Boolean isDefaultProfile) {
+        if (isDefaultProfile) {
+            deleteImage(currentProfileImage);
+            return S3Constants.DEFAULT_PROFILE_IMAGE;
+        }
+
+        if (newProfileImage.isEmpty()) {
+            return currentProfileImage;
+        }
+
+        deleteImage(currentProfileImage);
+        return uploadImage(newProfileImage);
+    }
+
     public void deleteImage(String imageUrl) {
-        String fileName = S3FileUtils.extractFileNameFromUrl(imageUrl);
-        amazonS3Client.deleteObject(bucketName, fileName);
+        if (!Objects.equals(S3Constants.DEFAULT_PROFILE_IMAGE, imageUrl)) {
+            String fileName = S3FileUtils.extractFileNameFromUrl(imageUrl);
+            amazonS3Client.deleteObject(bucketName, fileName);
+        }
     }
 
     private void uploadToS3(MultipartFile file, String fileName) {
